Start reorganizing the 15m tool to share functionality between modes. Extract common stuff out from the default isochrone-from-1-start viewer.
